+++ hyperfit/src/main/java/org/hyperfit/resource/controls/link/HyperLink.java
@@ -90,7 +90,7 @@ public abstract class HyperLink {
 
     public RequestBuilder toRequestBuilder() {
 
-        BoringRequestBuilder builder = new BoringRequestBuilder().setUrl(this.getHref());
+        BoringRequestBuilder builder = new BoringRequestBuilder(this.getHref());
 
         if(!StringUtils.isEmpty(this.getType())){
             builder.addAcceptedContentType(this.getType());